requirements.txt
google_api_python_client==2.138.0
gspread==6.1.2
numpy==2.0.1
Pillow==10.4.0
protobuf==5.27.2
python-dotenv==1.0.1
streamlit==1.35.0
streamlit_option_menu==0.3.12
